requirements.txt
flask
requests
supabase
python-dotenv
pathlib
logging